+++ govcd/system_test.go
@@ -15,40 +15,116 @@ import (
 	"github.com/vmware/go-vcloud-director/v2/util"
 )
 
-// Tests System function GetOrgByName by checking if the org object
-// return has the same name as the one provided in the config file.
+// Tests System methods vcd.GetOrg* by checking if the org object
+// returned has the same name as the one provided in the config file.
 // Asserts an error if the names don't match or if the function returned
 // an error. Also tests an org that doesn't exist. Asserts an error
-// if the function finds it or if the error is not nil.
-func (vcd *TestVCD) Test_GetOrgByName(check *C) {
-	org, err := GetOrgByName(vcd.client, vcd.config.VCD.Org)
-	check.Assert(org, Not(Equals), Org{})
-	check.Assert(err, IsNil)
-	check.Assert(org.Org.Name, Equals, vcd.config.VCD.Org)
-	// Tests Org That doesn't exist
-	org, err = GetOrgByName(vcd.client, INVALID_NAME)
-	check.Assert(org, Equals, Org{})
+// if the function finds it or if the error is nil.
+// Repeats the same operations fot GetOrgById, GetOrgByNameOrId, GetOrg
+func (vcd *TestVCD) Test_GetOrgByNameOrId(check *C) {
+
+	// Test Get by name
+	orgName := vcd.config.VCD.Org
+	org1, err1 := vcd.client.GetOrgByName(orgName)
+	check.Assert(org1, NotNil)
+	check.Assert(err1, IsNil)
+	check.Assert(org1.Org.Name, Equals, orgName)
+	orgId := org1.Org.ID
+	// Tests Org that doesn't exist
+	org1, err1 = vcd.client.GetOrgByName(INVALID_NAME)
+	check.Assert(org1, IsNil)
 	// When we explicitly search for a non existing item, we expect the error to be not nil
-	check.Assert(err, NotNil)
+	check.Assert(err1, NotNil)
+	check.Assert(IsNotFound(err1), Equals, true)
+
+	// Test Get by ID
+	org2, err2 := vcd.client.GetOrgById(orgId)
+	check.Assert(org2, NotNil)
+	check.Assert(err2, IsNil)
+	check.Assert(org2.Org.Name, Equals, orgName)
+	check.Assert(org2.Org.ID, Equals, orgId)
+	org2, err2 = vcd.client.GetOrgById(invalidEntityId)
+	check.Assert(org2, IsNil)
+	check.Assert(err2, NotNil)
+	check.Assert(IsNotFound(err2), Equals, true)
+
+	// Test Get by name or ID using the ID
+	org3, err3 := vcd.client.GetOrgByNameOrId(orgId)
+	check.Assert(org3, NotNil)
+	check.Assert(err3, IsNil)
+	check.Assert(org3.Org.Name, Equals, orgName)
+	check.Assert(org3.Org.ID, Equals, orgId)
+	org3, err3 = vcd.client.GetOrgByNameOrId(invalidEntityId)
+	check.Assert(org3, IsNil)
+	check.Assert(err3, NotNil)
+	check.Assert(IsNotFound(err3), Equals, true)
+
+	// Test Get by name or ID using the name
+	org4, err4 := vcd.client.GetOrgByNameOrId(orgName)
+	check.Assert(org4, NotNil)
+	check.Assert(err4, IsNil)
+	check.Assert(org4.Org.Name, Equals, orgName)
+	check.Assert(org4.Org.ID, Equals, orgId)
+	org4, err4 = vcd.client.GetOrgByNameOrId(INVALID_NAME)
+	check.Assert(org4, IsNil)
+	check.Assert(err4, NotNil)
+	check.Assert(IsNotFound(err4), Equals, true)
 }
 
-// Tests System function GetAdminOrgByName by checking if the AdminOrg object
-// return has the same name as the one provided in the config file. Asserts
-// an error if the names don't match or if the function returned an error.
-// Also tests an org that doesn't exist. Asserts an error
-// if the function finds it or if the error is not nil.
-func (vcd *TestVCD) Test_GetAdminOrgByName(check *C) {
-	if vcd.skipAdminTests {
-		check.Skip(fmt.Sprintf(TestRequiresSysAdminPrivileges, check.TestName()))
-	}
-	org, err := GetAdminOrgByName(vcd.client, vcd.config.VCD.Org)
-	check.Assert(org, Not(Equals), AdminOrg{})
-	check.Assert(err, IsNil)
-	check.Assert(org.AdminOrg.Name, Equals, vcd.config.VCD.Org)
-	// Tests Org That doesn't exist
-	org, err = GetAdminOrgByName(vcd.client, INVALID_NAME)
-	check.Assert(org, Equals, AdminOrg{})
-	check.Assert(err, NotNil)
+// Tests System methods vcd.GetAdminOrg* by checking if the adminOrg object
+// returned has the same name as the one provided in the config file.
+// Asserts an error if the names don't match or if the function returned
+// an error. Also tests an admin org that doesn't exist. Asserts an error
+// if the function finds it or if the error is nil.
+// Repeats the same operations fot GetAdminOrgById, GetAdminOrgByNameOrId, GetAdminOrg
+func (vcd *TestVCD) Test_GetAdminOrgByNameOrId(check *C) {
+
+	// Test Get by name
+	adminOrgName := vcd.config.VCD.Org
+	adminOrg1, err1 := vcd.client.GetAdminOrgByName(adminOrgName)
+	check.Assert(adminOrg1, NotNil)
+	check.Assert(err1, IsNil)
+	check.Assert(adminOrg1.AdminOrg.Name, Equals, adminOrgName)
+	orgId := adminOrg1.AdminOrg.ID
+	// Tests Org that doesn't exist
+	adminOrg1, err1 = vcd.client.GetAdminOrgByName(INVALID_NAME)
+	check.Assert(adminOrg1, IsNil)
+	// When we explicitly search for a non existing item, we expect the error to be not nil
+	check.Assert(err1, NotNil)
+	check.Assert(IsNotFound(err1), Equals, true)
+
+	// Test Get by ID
+	adminOrg2, err2 := vcd.client.GetAdminOrgById(orgId)
+	check.Assert(adminOrg2, NotNil)
+	check.Assert(err2, IsNil)
+	check.Assert(adminOrg2.AdminOrg.Name, Equals, adminOrgName)
+	check.Assert(adminOrg2.AdminOrg.ID, Equals, orgId)
+	adminOrg2, err2 = vcd.client.GetAdminOrgById(invalidEntityId)
+	check.Assert(adminOrg2, IsNil)
+	check.Assert(err2, NotNil)
+	check.Assert(IsNotFound(err2), Equals, true)
+
+	// Test Get by name or ID using the ID
+	adminOrg3, err3 := vcd.client.GetAdminOrgByNameOrId(orgId)
+	check.Assert(adminOrg3, NotNil)
+	check.Assert(err3, IsNil)
+	check.Assert(adminOrg3.AdminOrg.Name, Equals, adminOrgName)
+	check.Assert(adminOrg3.AdminOrg.ID, Equals, orgId)
+	adminOrg3, err3 = vcd.client.GetAdminOrgByNameOrId(invalidEntityId)
+	check.Assert(adminOrg3, IsNil)
+	check.Assert(err3, NotNil)
+	check.Assert(IsNotFound(err3), Equals, true)
+
+	// Test Get by name or ID using the name
+	adminOrg4, err4 := vcd.client.GetAdminOrgByNameOrId(adminOrgName)
+	check.Assert(adminOrg4, NotNil)
+	check.Assert(err4, IsNil)
+	check.Assert(adminOrg4.AdminOrg.Name, Equals, adminOrgName)
+	check.Assert(adminOrg4.AdminOrg.ID, Equals, orgId)
+	adminOrg4, err4 = vcd.client.GetAdminOrgByNameOrId(INVALID_NAME)
+	check.Assert(adminOrg4, IsNil)
+	check.Assert(err4, NotNil)
+	check.Assert(IsNotFound(err4), Equals, true)
 }
 
 // Tests the creation of an org with general settings,
@@ -58,19 +134,26 @@ func (vcd *TestVCD) Test_CreateOrg(check *C) {
 	if vcd.skipAdminTests {
 		check.Skip(fmt.Sprintf(TestRequiresSysAdminPrivileges, check.TestName()))
 	}
-	org, _ := GetAdminOrgByName(vcd.client, TestCreateOrg)
-	if org != (AdminOrg{}) {
-		err := org.Delete(true, true)
-		check.Assert(err, IsNil)
+
+	type testOrgData struct {
+		name                     string
+		enabled                  bool
+		canPublishCatalogs       bool
+		deployedVmQuota          int
+		storedVmQuota            int
+		delayAfterPowerOnSeconds int
+		fullData                 bool
 	}
-	settings := &types.OrgSettings{
-		OrgGeneralSettings: &types.OrgGeneralSettings{
-			CanPublishCatalogs:       true,
-			DeployedVMQuota:          10,
-			StoredVMQuota:            10,
-			UseServerBootSequence:    true,
-			DelayAfterPowerOnSeconds: 3,
-		},
+	var orgList = []testOrgData{
+		{"org1", true, false, 0, 0, 0, true},
+		{"org2", true, true, 0, 0, 1, false},
+		{"org3", false, false, 1, 1, 3, true},
+		{"org4", true, true, 10, 10, 10, false},
+		{"org5", false, true, 100, 100, 100, false},
+	}
+
+	fullSettings := &types.OrgSettings{
+		OrgGeneralSettings: &types.OrgGeneralSettings{},
 		OrgVAppTemplateSettings: &types.VAppTemplateLeaseSettings{
 			DeleteOnStorageLeaseExpiration: true,
 			StorageLeaseSeconds:            10,
@@ -85,23 +168,46 @@ func (vcd *TestVCD) Test_CreateOrg(check *C) {
 			OrgLdapMode: "NONE",
 		},
 	}
-	task, err := CreateOrg(vcd.client, TestCreateOrg, TestCreateOrg, TestCreateOrg, settings, true)
-	check.Assert(err, IsNil)
-	// After a successful creation, the entity is added to the cleanup list.
-	// If something fails after this point, the entity will be removed
-	AddToCleanupList(TestCreateOrg, "org", "", "TestCreateOrg")
-	err = task.WaitTaskCompletion()
-	check.Assert(err, IsNil)
-	// fetch newly created org
-	org, err = GetAdminOrgByName(vcd.client, TestCreateOrg)
-	check.Assert(org, Not(Equals), AdminOrg{})
-	check.Assert(err, IsNil)
-	check.Assert(org.AdminOrg.Name, Equals, TestCreateOrg)
-	check.Assert(org.AdminOrg.Description, Equals, TestCreateOrg)
-	// Delete, with force and recursive true
-	err = org.Delete(true, true)
-	check.Assert(err, IsNil)
-	doesOrgExist(check, vcd)
+	for _, od := range orgList {
+		var settings *types.OrgSettings
+		if od.fullData {
+			settings = fullSettings
+		} else {
+			settings = &types.OrgSettings{
+				OrgGeneralSettings: &types.OrgGeneralSettings{},
+			}
+		}
+		orgName := TestCreateOrg + "_" + od.name
+
+		fmt.Printf("# org %s (enabled: %v - catalogs: %v [%d %d])\n", orgName, od.enabled, od.canPublishCatalogs, od.storedVmQuota, od.deployedVmQuota)
+		settings.OrgGeneralSettings.CanPublishCatalogs = od.canPublishCatalogs
+		settings.OrgGeneralSettings.DeployedVMQuota = od.deployedVmQuota
+		settings.OrgGeneralSettings.StoredVMQuota = od.storedVmQuota
+		settings.OrgGeneralSettings.DelayAfterPowerOnSeconds = od.delayAfterPowerOnSeconds
+		task, err := CreateOrg(vcd.client, orgName, TestCreateOrg, TestCreateOrg, settings, od.enabled)
+		check.Assert(err, IsNil)
+		// After a successful creation, the entity is added to the cleanup list.
+		// If something fails after this point, the entity will be removed
+		AddToCleanupList(orgName, "org", "", "TestCreateOrg")
+		err = task.WaitTaskCompletion()
+		check.Assert(err, IsNil)
+		// fetch newly created org
+		adminOrg, err := vcd.client.GetAdminOrgByName(orgName)
+		check.Assert(err, IsNil)
+		check.Assert(adminOrg, NotNil)
+		check.Assert(adminOrg.AdminOrg.Name, Equals, orgName)
+		check.Assert(adminOrg.AdminOrg.Description, Equals, TestCreateOrg)
+		check.Assert(adminOrg.AdminOrg.IsEnabled, Equals, od.enabled)
+
+		check.Assert(adminOrg.AdminOrg.OrgSettings.OrgGeneralSettings.CanPublishCatalogs, Equals, od.canPublishCatalogs)
+		check.Assert(adminOrg.AdminOrg.OrgSettings.OrgGeneralSettings.DeployedVMQuota, Equals, od.deployedVmQuota)
+		check.Assert(adminOrg.AdminOrg.OrgSettings.OrgGeneralSettings.StoredVMQuota, Equals, od.storedVmQuota)
+		check.Assert(adminOrg.AdminOrg.OrgSettings.OrgGeneralSettings.DelayAfterPowerOnSeconds, Equals, od.delayAfterPowerOnSeconds)
+		// Delete, with force and recursive true
+		err = adminOrg.Delete(true, true)
+		check.Assert(err, IsNil)
+		doesOrgExist(check, vcd)
+	}
 }
 
 func (vcd *TestVCD) Test_CreateDeleteEdgeGateway(check *C) {
@@ -197,9 +303,9 @@ func (vcd *TestVCD) Test_GetNetworkPoolByHREF(check *C) {
 
 	fmt.Printf("Running: %s\n", check.TestName())
 
-	adminOrg, err := GetAdminOrgByName(vcd.client, vcd.config.VCD.Org)
-	check.Assert(adminOrg, Not(Equals), AdminOrg{})
+	adminOrg, err := vcd.client.GetAdminOrgByName(vcd.config.VCD.Org)
 	check.Assert(err, IsNil)
+	check.Assert(adminOrg, NotNil)
 
 	adminVdc, err := adminOrg.GetAdminVdcByName(vcd.config.VCD.Vdc)
 	check.Assert(adminVdc, Not(Equals), AdminVdc{})
@@ -210,11 +316,6 @@ func (vcd *TestVCD) Test_GetNetworkPoolByHREF(check *C) {
 	check.Assert(err, IsNil)
 	check.Assert(foundNetworkPool, Not(Equals), types.VMWNetworkPool{})
 }
-
-// longer than the 128 characters so nothing can be named this
-var INVALID_NAME = `*******************************************INVALID
-					****************************************************
-					************************`
 
 func init() {
 	testingTags["system"] = "system_test.go"
@@ -232,4 +333,76 @@ func (vcd *TestVCD) Test_QueryOrgVdcNetworkByName(check *C) {
 	check.Assert(len(orgVdcNetwork), Not(Equals), 0)
 	check.Assert(orgVdcNetwork[0].Name, Equals, vcd.config.VCD.Network.Net1)
 	check.Assert(orgVdcNetwork[0].ConnectedTo, Equals, vcd.config.VCD.EdgeGateway)
+}
+
+func (vcd *TestVCD) Test_QueryProviderVdcEntities(check *C) {
+	providerVdcName := vcd.config.VCD.ProviderVdc.Name
+	networkPoolName := vcd.config.VCD.ProviderVdc.NetworkPool
+	storageProfileName := vcd.config.VCD.ProviderVdc.StorageProfile
+	if providerVdcName == "" {
+		check.Skip("Skipping Provider VDC query: no provider VDC was given")
+	}
+	providerVdcs, err := vcd.client.QueryProviderVdcs()
+	check.Assert(err, IsNil)
+	check.Assert(len(providerVdcName) > 0, Equals, true)
+
+	providerFound := false
+	for _, providerVdc := range providerVdcs {
+		if providerVdcName == providerVdc.Name {
+			providerFound = true
+		}
+
+		if testVerbose {
+			fmt.Printf("PVDC %s\n", providerVdc.Name)
+			fmt.Printf("\t href    %s\n", providerVdc.HREF)
+			fmt.Printf("\t status  %s\n", providerVdc.Status)
+			fmt.Printf("\t enabled %v\n", providerVdc.IsEnabled)
+			fmt.Println("")
+		}
+	}
+	check.Assert(providerFound, Equals, true)
+
+	if networkPoolName == "" {
+		check.Skip("Skipping Network pool query: no network pool was given")
+	}
+	netPools, err := vcd.client.QueryNetworkPools()
+	check.Assert(err, IsNil)
+	check.Assert(len(netPools) > 0, Equals, true)
+	networkPoolFound := false
+	for _, networkPool := range netPools {
+		if networkPoolName == networkPool.Name {
+			networkPoolFound = true
+		}
+		if testVerbose {
+			fmt.Printf("NP %s\n", networkPool.Name)
+			fmt.Printf("\t href %s\n", networkPool.HREF)
+			fmt.Printf("\t type %v\n", networkPool.NetworkPoolType)
+			fmt.Println("")
+		}
+	}
+	check.Assert(networkPoolFound, Equals, true)
+
+	if storageProfileName == "" {
+		check.Skip("Skipping storage profile query: no storage profile was given")
+	}
+	storageProfiles, err := vcd.client.QueryProviderVdcStorageProfiles()
+	check.Assert(err, IsNil)
+	check.Assert(len(storageProfiles) > 0, Equals, true)
+	storageProfileFound := false
+	for _, sp := range storageProfiles {
+		if storageProfileName == sp.Name {
+			storageProfileFound = true
+		}
+		if testVerbose {
+			fmt.Printf("SP %s\n", sp.Name)
+			fmt.Printf("\t enabled     %12v\n", sp.IsEnabled)
+			fmt.Printf("\t storage     %12d\n", sp.StorageTotalMB)
+			fmt.Printf("\t provisioned %12d\n", sp.StorageProvisionedMB)
+			fmt.Printf("\t requested   %12d\n", sp.StorageRequestedMB)
+			fmt.Printf("\t used        %12d\n", sp.StorageUsedMB)
+			fmt.Println("")
+		}
+	}
+	check.Assert(storageProfileFound, Equals, true)
+
 }